GENTAX-AI/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
langchain==0.1.20
langchain-community==0.0.38
langchain-groq==0.1.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pydantic==1.10.13





